+++ tools/lsp/preview/connector.rs
@@ -14,8 +14,6 @@ pub use native::*;
 pub use crate::common::SwitchableLspToPreview;
 #[cfg(feature = "preview-remote")]
 pub mod remote;
-#[cfg(feature = "preview-remote")]
-pub use remote::*;
 
 use crate::preview;
